task_manager/task_manager/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
mysql-connector-python==8.4.0
mysqlclient==2.2.4
sqlparse==0.5.0
